review: login page를 기능과 레이아웃으로 코드 분할
