+++ src/app/components/common/Popup.tsx
@@ -2,7 +2,7 @@ import { popupContents, popupWrapper } from "@/app/styles/common/popup.css";
 import { paragraph2, semiBold } from "@/app/styles/font.css";
 import { flexSprinklesFc } from "./utils/flex";
 
-const Popup = ({ text, children }: { text: string; children: React.ReactNode }) => {
+const Popup = ({ text, children }: { text: null | string; children: React.ReactNode }) => {
   return (
     <section className={popupWrapper}>
       <article className={`${popupContents} ${flexSprinklesFc({ flexDirection: "column", gap: "24px" })}`}>
