+++ src/app/login/hook/index.ts
@@ -1,0 +1,1 @@
+export { useAuth } from "./useAuth";
